Detach JSON services from Learning
Preparation du deplacement des services JSON vers Norm
- objectif:
  - reutilisation du parser JSON dans Norm pour le pilotage par fichiers scenarios et json
  - extraction ce qui n'est pas specifique a Learning: tout, sauf JSONFile
  - nettoyage et mise en conformite avec la bibliotheque Norm

Impacts principaux
- suppression des references au type Continuous
- renommages:
  - KWTextService en TextService
  - KWIntVectorSorter en IntVectorSorter
- TextService:
  - enrichissement de la classe en rappatriant des methdoes de services generique provenant de JSONFile
  - maintenant, classe ancetre de
    - JSONFile
    - JSONTokenizer
    - KWTextTokenizer
- IntVectorSorter: nouveau fichier dedie au classes IntPairVector et IntVectorSorter, extrait de KWSortableIndex

Impacts detailles
- Supression des references à Continuous dans les services JSON
  - JsonLex.lex: utilisation desormais de strtod pour parser les doubles
  - supression des include inutiles
    - JSONTokenizer.h: plus d'inclusion de KWContinuous.h
    - KWTextService.h: plus d'inclusion de KWTokenFrequency.h
- JSONTokenizer
  - supression de la dependance au type Continuous
    - supression des methodes ReadContinuousValue et ReadKeyContinuousValue
    - methode interne DoubleToInt, pour se passer de KWContinuous::ContinuousToInt
  - impacts dans CCCoclusteringReport, lecture de fichier json
- TextServive:
  - issue d'un renommage de KWTextService en TextService
  - impacts mineurs lies au renommage
    - KWClass
    - KDTextFeatureConstruction
    - KDTextTokenSampleCollectionTask
    - KWTokenFrequency
    - JSONLex.lex
    - KWDRRuleLIbrary/KWDRTextualAnalysis
    - KWTest/Test
  - recupere l'essentiel des methodes de service provenant de JSONFile
  - maintenant classe ancetre de:
    - JSONFile
    - JSONTokenizer
    - KWTextTokenizer
- IntVectorSorter
  - nouveau fichier dedie au classes IntPairVector et IntVectorSorter
  - issu du renommage de KWIntVectorSorter en IntVectorSorter
    - impacts:
      - KWDatabaseFormatDetector
      - TextService
  - issu de KWSortableIndex, dont il est extraiut
  - usages dans KWDatabaseFormatDetector et TextService
