+++ src/Learning/KWData/JSONFile.h
@@ -12,12 +12,13 @@ class JSONFile;
 #include "FileService.h"
 #include "MemoryStatsManager.h"
 #include "PLRemoteFileService.h"
-#include "KWSortableIndex.h"
-#include "KWTextService.h"
+#include "TextService.h"
 
 //////////////////////////////////////////
 // Fichier JSON
-class JSONFile : public Object
+// Service specifique aux rapports de modelisation, avec gestion du type Continuous
+// et de l'encodage des fichier par une balise dediee en fin de fichier JSON
+class JSONFile : public TextService
 {
 public:
 	// Constructeur
@@ -115,25 +116,6 @@ public:
 	static void SetVerboseMode(boolean bValue);
 	static boolean GetVerboseMode();
 
-	///////////////////////////////////////////////////////////////////////////////
-	// Gestion de l'encodage des chaines ansi en utf8, le format supporte par json
-	// Les caracteres ansi 128 a 255 sont encodes avec iso-8859-1/windows-1252
-	// L'encodage dans le json se fait avec le caractere d'echapement \uHHHH
-
-	// Encodage d'un chaine de caracteres C au format json, sans les double-quotes de debut et fin
-	static void CStringToJsonString(const ALString& sCString, ALString& sJsonString);
-
-	// Encodage d'une chaine de caracteres C au format ansi en re-encodant les caracteres utf8
-	// endodes avec iso-8859-1/windows-1252 vers ansi
-	static void CStringToCAnsiString(const ALString& sCString, ALString& sCAnsiString);
-
-	// Conversion d'un caractere ansi windows-1252 vers un caractere unicode au format hexa
-	static void Windows1252ToUnicodeHex(int nAnsiCode, ALString& sUnicodeHexChars);
-
-	// Conversion d'un caractere unicode au format hexa vers le code ansi windows-1252
-	// Renvoie un code ansi entre 0 et 255 si ok, -1 sinon
-	static int UnicodeHexToWindows1252(const ALString& sUnicodeHexChars);
-
 	////////////////////////////////////////////////////
 	// Divers
 
@@ -181,46 +163,6 @@ protected:
 	// Exploitation des stats d'encodage: memorisation dans le fichier json et message utilisateur
 	void ExploitEncodingStats();
 
-	//////////////////////////////////////////////////////////////////////////////////////
-	// Gestion des encodages windows-1252 vers unicode et UFT8
-
-	// Code d'une chaine de caracteres hexa encodant un caractere unicode ou utf8,
-	// pour une longueur d'au plus trois byte correspondant aux besoins d'encodage de la page de code Windows-1252
-	static int GetHexStringCode(const ALString& sHexString);
-
-	// Conversion d'un caractere ansi windows-1252 vers un caractere utf8 au format hexa
-	static void Windows1252ToUtf8Hex(int nAnsiCode, ALString& sUtf8HexChars);
-
-	// Recherche du code ansi d'un code de caracteres utf8 dans la page de code Windows-1252
-	// Renvoie un code ansi entre 0 et 255 si ok, -1 sinon
-	static int Windows1252Utf8CodeToWindows1252(int nWindows1252Utf8Code);
-
-	// Initialisation de l'ensemble des structures d'encodages
-	static void InitializeEncodingStructures();
-
-	// Indique que l'initialisation est effectuee
-	static boolean AreEncodingStructuresInitialized();
-
-	// Verification de l'ensemble des structures d'encodages
-	static boolean CheckEncodingStructures();
-
-	// Initialisation de la table de transcodage entre les caracteres asci Windows-1252 et les caracteres unicode ou
-	// utf8 Les caracteres speciaux 0x00 a 0x1F sont encodes en 0x00HH (4 caractere hexa en unicode) Les caracteres
-	// ansi 0x20 a 0x7F sont encodes tels quels avec un seul caractere Les caracteres speciaux 0x80 a 0x9F sont
-	// encodes de facon speciale pour avoir le meme caractere imprimable qu'avec l'encodage windows-1252 (4
-	// caractere hexa en unicode). Cf. https://www.i18nqa.com/debug/table-iso8859-1-vs-windows-1252.html
-	// 	   https://www.charset.org/utf-8
-	// Les caracteres latin etendus speciaux 0xA0 a 0xEF sont encodes en 0x00HH (4 caractere hexa)
-	static void InitializeWindows1252UnicodeHexEncoding();
-	static void InitializeWindows1252Utf8HexEncoding();
-
-	// Nombre max de bytes de l'encodage windows-1252
-	const static int nWindows1252EncodingMaxByteNumber = 3;
-
-	// Initialisation des structure de decodage des caracteres Windows-1252, pour retrouver leur index en fonction
-	// de leur code utf8
-	static void InitializeWindows1252Utf8DecodingStructures();
-
 	///////////////////////////////////////////////////////////////////////////////////
 	// Variable d'instance
 
@@ -240,9 +182,9 @@ protected:
 	// Parametrage des cles en mode camel case
 	boolean bCamelCaseKeys;
 
-	// Vecteurs de collecte de stats pour les caracteres ansi etendus (128 a 255) encodes soit en unicode, soit
-	// directement presents avec leur encodage. Les vecteurs sont de taille 256, mais seul les index 128 a 255 sont
-	// collectes
+	// Vecteurs de collecte de stats pour les caracteres ansi etendus (128 a 255) encodes soit en unicode,
+	// soit directement presents avec leur encodage.
+	// Les vecteurs sont de taille 256, mais seul les index 128 a 255 sont collectes
 	LongintVector lvWindows1252AnsiCharNumbers;
 	LongintVector lvWindows1252Utf8CharNumbers;
 
@@ -255,23 +197,6 @@ protected:
 	// caracteres
 	ALString sStringBuffer;
 
-	///////////////////////////////////////////////////////////////////////////////////
-	// Gestion des encodage windows-1252 vers l'unicode et l'UTF8, en caracteres hexa
-
 	// Mode verbeux
 	static boolean bVerboseMode;
-
-	// Table de transcodage de taille 256 entre les caracteres asci windows-1252 et les caracteres unicode, ou utf8,
-	// sous forme de chaines de caracteres en hexa (0-9 et A-F)
-	static StringVector svWindows1252UnicodeHexEncoding;
-	static StringVector svWindows1252Utf8HexEncoding;
-
-	// Vecteur des codes UTF8 pour les 32 caracteres de controles Windows-1252 (128 a 159), et leur index
-	static KWIntVectorSorter ivsWindows1252ControlCharUtf8CodeSorter;
-
-	// Instance statique de JSONFile, permettant de forcer l'initialisation des structure d'encodage une fois
-	// pour toute lors de l'appel du constructeur de cette instance
-	// Ne pas declarer d'autre instances statiques de JSONFile, par exemple via d'autre classes, sinon cela pose
-	// des probleme de memoire non liberee non diagnostique par les outils de getsion de la memoire
-	static JSONFile jsonFileGlobalInitializer;
 };
