+++ src/Learning/MODL_Coclustering/CCCoclusteringReport.cpp
@@ -612,8 +612,8 @@ boolean CCCoclusteringReport::ReadDimensionSummaries(CCHierarchicalDataGrid* coc
 		// Valeur min et max dans le cas numerique
 		if (bOk and nAttributeType == KWType::Continuous)
 		{
-			bOk = bOk and JSONTokenizer::ReadKeyContinuousValue("min", false, cMin, bIsEnd);
-			bOk = bOk and JSONTokenizer::ReadKeyContinuousValue("max", false, cMax, bIsEnd);
+			bOk = bOk and JSONTokenizer::ReadKeyDoubleValue("min", false, cMin, bIsEnd);
+			bOk = bOk and JSONTokenizer::ReadKeyDoubleValue("max", false, cMax, bIsEnd);
 		}
 
 		// Initialisation de la grille ou ajout d'un attribut
@@ -812,8 +812,8 @@ boolean CCCoclusteringReport::ReadInnerAttributesDimensionSummaries(KWDGAttribut
 		// Valeur min et max dans le cas numerique
 		if (bOk and nAttributeType == KWType::Continuous)
 		{
-			bOk = bOk and JSONTokenizer::ReadKeyContinuousValue("min", false, cMin, bIsEnd);
-			bOk = bOk and JSONTokenizer::ReadKeyContinuousValue("max", false, cMax, bIsEnd);
+			bOk = bOk and JSONTokenizer::ReadKeyDoubleValue("min", false, cMin, bIsEnd);
+			bOk = bOk and JSONTokenizer::ReadKeyDoubleValue("max", false, cMax, bIsEnd);
 		}
 
 		// Creation et specification de l'attribut interne
@@ -1318,7 +1318,7 @@ boolean CCCoclusteringReport::ReadInterval(KWDGAttribute* dgAttribute, KWDGPart*
 
 			// Lecture de la borne sup
 			bOk = bOk and JSONTokenizer::ReadExpectedToken(',');
-			bOk = bOk and JSONTokenizer::ReadContinuousValue(false, cUpperBound);
+			bOk = bOk and JSONTokenizer::ReadDoubleValue(false, cUpperBound);
 			bOk = bOk and JSONTokenizer::ReadExpectedToken(']');
 		}
 	}
